refactoring fast median filter interface
